LayoutTests/storage/indexeddb/transaction-read-only-expected.txt

Test read-only transactions in IndexedDB.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitIndexedDB' in window is true
PASS webkitIndexedDB == null is false
webkitIndexedDB.open('name')
PASS 'onsuccess' in result is true
PASS 'onerror' in result is true
PASS 'readyState' in result is true
An event should fire shortly...

openSuccess():
Success event fired:
PASS 'result' in event is true
PASS 'code' in event is false
PASS 'message' in event is false
PASS 'source' in event is true
PASS event.source != null is true
PASS 'onsuccess' in event.target is true
PASS 'onerror' in event.target is true
PASS 'readyState' in event.target is true
PASS event.target.readyState is event.target.DONE

db = event.result
result = db.setVersion('version 1')
Success event fired:
PASS 'result' in event is true
PASS 'code' in event is false
PASS 'message' in event is false
PASS 'source' in event is true
PASS event.source != null is true
PASS 'onsuccess' in event.target is true
PASS 'onerror' in event.target is true
PASS 'readyState' in event.target is true
PASS event.target.readyState is event.target.DONE

Deleted all object stores.
store = db.createObjectStore('store')
store.put('x', 'y')
trans = db.transaction()
Expecting exception from trans.objectStore('store').put('a', 'b')
PASS code is webkitIDBDatabaseException.READ_ONLY_ERR
trans = db.transaction()
Expecting exception from trans.objectStore('store').delete('x')
PASS code is webkitIDBDatabaseException.READ_ONLY_ERR
trans = db.transaction()
cur = trans.objectStore('store').openCursor()
PASS !event.result is false
Expecting exception from event.result.delete()
PASS code is webkitIDBDatabaseException.READ_ONLY_ERR
PASS successfullyParsed is true

TEST COMPLETE
